feature: refactor table
