+++ src/routes/table.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import ViewBeerLocationsTable from "@feature/table";
+import ViewBeerLocationsTable from "@feature/table/view-beer-locations-table";
 
 export const Route = createFileRoute("/table")({
 	component: ViewBeerLocationsTable,
